Move all dependencies in project/Deps.scala
