+++ project/Deps.scala
@@ -1,23 +1,33 @@
 
+import org.portablescala.sbtplatformdeps.PlatformDepsPlugin.autoImport._
 import sbt._
 import sbt.Def.setting
 import sbt.Keys._
 
 object Deps {
 
-  def catsCore = "org.typelevel" %% "cats-core" % "1.6.0"
-  def fastParse = "com.lihaoyi" %% "fastparse" % SharedVersions.fastParse
-  def jsoup = "org.jsoup" % "jsoup" % "1.12.1"
-  def scalaXml = "org.scala-lang.modules" %% "scala-xml" % "1.2.0"
-  def scalazConcurrent = "org.scalaz" %% "scalaz-concurrent" % SharedVersions.scalaz
+  private object versions {
+    def argonautShapeless = "1.2.0-M11"
+    def fastParse = "2.1.2"
+    def http4s = "0.18.17"
+    def scalaz = "7.2.27"
+  }
+
+  def argonautShapeless = "com.github.alexarchambault" %% "argonaut-shapeless_6.2" % versions.argonautShapeless
   def caseApp = "com.github.alexarchambault" %% "case-app" % "2.0.0-M8"
+  def catsCore = "org.typelevel" %% "cats-core" % "1.6.0"
+  def dockerClient = "com.spotify" % "docker-client" % "8.16.0"
+  def emoji = "com.lightbend" %% "emoji" % "1.2.1"
+  def fastParse = "com.lihaoyi" %% "fastparse" % versions.fastParse
+  def http4sBlazeServer = "org.http4s" %% "http4s-blaze-server" % versions.http4s
+  def http4sDsl = "org.http4s" %% "http4s-dsl" % versions.http4s
+  def jsoup = "org.jsoup" % "jsoup" % "1.12.1"
+  def junit = "junit" % "junit" % "4.12"
+  def logbackClassic = "ch.qos.logback" % "logback-classic" % "1.2.3"
+  def mavenModel = "org.apache.maven" % "maven-model" % "3.6.1"
   // staying before 3.14.0 to the URLStreamHandlerFactory implementation
   def okhttpUrlConnection = "com.squareup.okhttp3" % "okhttp-urlconnection" % "3.13.1"
-  def argonautShapeless = "com.github.alexarchambault" %% "argonaut-shapeless_6.2" % SharedVersions.argonautShapeless
-  def scalatest = "org.scalatest" %% "scalatest" % "3.0.7"
-  def junit = "junit" % "junit" % "4.12"
-  def dockerClient = "com.spotify" % "docker-client" % "8.16.0"
-
+  def slf4JNop = "org.slf4j" % "slf4j-nop" % "1.7.26"
   def scalaAsync = Def.setting {
     val sv = scalaVersion.value
     val ver =
@@ -25,11 +35,27 @@ object Deps {
       else "0.10.0"
     "org.scala-lang.modules" %% "scala-async" % ver
   }
-
   def scalaNativeTools03 = "org.scala-native" %% "tools" % "0.3.9"
   def scalaNativeTools040M2 = "org.scala-native" %% "tools" % "0.4.0-M2"
+  def scalaReflect = setting("org.scala-lang" % "scala-reflect" % scalaVersion.value)
+  def scalatest = "org.scalatest" %% "scalatest" % "3.0.7"
+  def scalaXml = "org.scala-lang.modules" %% "scala-xml" % "1.2.0"
+  def scalazConcurrent = "org.scalaz" %% "scalaz-concurrent" % versions.scalaz
 
-  def slf4JNop = "org.slf4j" % "slf4j-nop" % "1.7.26"
+  object cross {
+    // The setting / .value hoop-and-loop is necessary because of the expansion of the %%% macro, which references
+    // other settings.
 
-  def scalaReflect = setting("org.scala-lang" % "scala-reflect" % scalaVersion.value) 
+    def argonautShapeless = setting("com.github.alexarchambault" %%% "argonaut-shapeless_6.2" % versions.argonautShapeless)
+    def catsEffect = setting("org.typelevel" %%% "cats-effect" % "1.2.0")
+    def fastParse = setting("com.lihaoyi" %%% "fastparse" % versions.fastParse)
+    def scalaJsDom = setting("org.scala-js" %%% "scalajs-dom" % "0.9.7")
+    def scalaJsJquery = setting("be.doeraene" %%% "scalajs-jquery" % "0.9.5")
+    def scalaJsReact = setting("com.github.japgolly.scalajs-react" %%% "core" % "1.3.1")
+    def scalazCore = setting("org.scalaz" %%% "scalaz-core" % versions.scalaz)
+    def utest = setting("com.lihaoyi" %%% "utest" % "0.6.7")
+  }
+
+  def proguardVersion = "6.0.3"
+
 }
