--- project/SharedVersions.scala
@@ -1,9 +1,0 @@
-
-object SharedVersions {
-
-  def argonautShapeless = "1.2.0-M11"
-  def fastParse = "2.1.2"
-  def proguard = "6.0.3"
-  def scalaz = "7.2.27"
-
-}
